[X8] got basic build working
crashes due to some NIB issue

+++ Soulver/SVRDocumentViewController.m
@@ -149,7 +149,7 @@ NSString *SVRDocumentViewControllerUnsolvedPasteboardType = @"com.saturdayapps.n
     case 23: return [@"10" stringByAppendingString:[NSString SVR_logRawString]];
     case 13: return nil; // Backspace key
   }
-  XPLogRaise2(@"<%@> Button with unknown tag: %ld", self, tag);
+  XPLogRaise2(@"<%@> Button with unknown tag: %d", self, (int)tag);
   return nil;
 }
